Finished simulate function. Need to adjust tests.
